AI_API/requirements.txt
setuptools
openai-whisper
pyannote.audio
soundfile
speechbrain==0.5.16
numpy==1.26.4
requests
python-dotenv
onnxruntime
torch
torchaudio
fastapi
uvicorn
pydub